requirements.txt
streamlit
chromadb
scikit-learn
hdbscan
numpy
pandas
matplotlib
plotly
umap-learn
sentence-transformers
